requirements.txt
asgiref==3.3.1
Django==3.1.6
django-webpack-loader==0.7.0
pio==0.0.3
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.1
djangorestframework==3.12.4
